frontend Settings: Create separate files for each component
